Fixed getNumMixed to return long

+++ src/main/java/hivemall/classifier/BinaryOnlineClassifierUDTF.java
@@ -300,7 +300,7 @@ public abstract class BinaryOnlineClassifierUDTF extends LearnerBaseUDTF {
                     numForwarded++;
                 }
             }
-            int numMixed = model.getNumMixed();
+            long numMixed = model.getNumMixed();
             this.model = null;
             logger.info("Trained a prediction model using " + count
                     + " training examples"
